Implement improved monitoring workflow with logical sequence
- Replace constant 7-second updates with logical workflow:
  * Show system status once
  * Process candidates if any exist (with progress tracking)
  * Show updated status after processing
  * Wait 1 minute before next cycle

- Add new monitoring methods:
  * _display_system_status() - clean status display
  * _process_candidates_with_progress() - batch processing with progress
  * _countdown_wait() - interruptible countdown timer

- Benefits:
  * Reduces system load (no constant polling)
  * Better user experience with clear workflow
  * Progress tracking during candidate processing
  * Logical sequence of operations

Fixes excessive monitoring updates issue.
